+++ spec/spec_helper.rb
@@ -11,4 +11,5 @@ RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
   config.mock_with :rspec
+  config.color = true
 end