+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+* Remove unused `pry` dependency
+* Update versions of dependencies
+
 ## 0.4.1
 
 * Improve notifications. Via #30